feat!: Complete migration to Cobra docs: update README for use with Cobra
