--- cobra/.gitignore
@@ -1,2 +1,0 @@
-config.toml
-*_markdown/
