--- cobra/config.toml.example
@@ -1,5 +1,0 @@
-[[destinations]]
-name = 'blogger'
-blog_url = 'https://example.com'
-type = 'blogger'
-
